fix: rename component files for clarity
